text-fieldのcssを簡素化した。また挙動を一部修正
