+++ packages/ui/src/app.css
@@ -2,6 +2,3 @@
   font-family: 'Helvetica Neue', Arial, 'Hiragino Kaku Gothic ProN',
     'Hiragino Sans', Meiryo, sans-serif;
 }
-.flex {
-  display: flex;
-}
