+++ packages/ui/src/components/server/text-field/filled.module.scss
@@ -1,0 +1,79 @@
+@use '@theme/css/color/color.scss' as *;
+@use '@theme/css/margin/margin.scss' as *;
+@use '@layer/layer.scss' as layer;
+
+// focusしていないときの枠線の色
+$blur-outline-color: $accents4;
+// focusしているときの枠線の色
+$focus-outline-color: $suoh;
+// focusしていないときの文字色
+$blur-color: $accents3;
+// focusしているときの文字色
+$focus-color: $suoh;
+// placeholderのアニメーションにかかる時間
+$transition-time: 0.3s;
+// z-indexの基準 = 0
+$base-z-index: layer.$earth;
+
+// 外枠
+.text-field {
+  position: relative;
+  height: 52px;
+  background-color: $accents2;
+  border-bottom: $blur-outline-color solid 1px;
+
+  // focusしたときの枠線のアニメーション
+  &::after {
+    border-bottom: 1px solid $suoh;
+    left: 0px;
+    bottom: -1px;
+    content: '';
+    position: absolute;
+    right: 0px;
+    transition: transform 200ms cubic-bezier(0, 0, 0.2, 1) 0ms;
+    pointer-events: none;
+  }
+  &:has(.input:focus)::after {
+    transform: scaleX(1);
+  }
+
+  &:not(:has(.input:focus))::after {
+    transform: scaleX(0);
+  }
+}
+
+// デフォルトのplaceholder
+.placeholder {
+  position: absolute;
+  margin: 0 12px;
+  transform-origin: 0;
+  transition: all $transition-time;
+  pointer-events: none;
+  color: $blur-color;
+
+  // placeholderを上下中央寄せする
+  top: 50%;
+  transform: translateY(-50%);
+  -webkit-transform: translateY(-50%);
+  -ms-transform: translateY(-50%);
+}
+
+// focusしているときおよび文字入力があるときのplaceholder
+.input:focus + .placeholder,
+.input:not(:placeholder-shown) + .placeholder {
+  transform: scale(0.8);
+  top: 0;
+  transition: all $transition-time;
+  pointer-events: none;
+  color: $focus-color;
+}
+
+// デフォルトのinput
+.input {
+  position: absolute;
+  bottom: 0;
+  padding: 13px 12px 0 12px;
+  width: 100%;
+  height: 100%;
+  z-index: $base-z-index;
+}
