+++ packages/ui/src/components/server/text-field/index.ts
@@ -1,1 +1,1 @@
-export { TextField } from './TextField'
+export { TextField } from './TextFiled'
