--- packages/ui/src/components/server/text-field/variant/outlined-text-field/outlined-text-field.module.scss
@@ -1,100 +1,0 @@
-@use '@theme/css/color/color.scss' as color;
-@use '@theme/css/font/index.scss' as font;
-@use '@theme/css/margin/margin.scss' as *;
-@use '@layer/layer.scss' as layer;
-
-@mixin use-outlined-text-field() {
-  $blur-outline-color: color.$accents4;
-  $focus-outline-color: color.$suoh;
-  $focus-color: color.$suoh;
-  $blur-color: color.$accents3;
-
-  $f-transition-time: 0.1s;
-  $b-transition-time: 0.1s;
-
-  $base-z-index: layer.$earth;
-
-  $blur-placeholder-z-index: $base-z-index;
-  $input-z-index: calc($base-z-index + 10);
-  $placeholder-background-z-index: calc($base-z-index + 20);
-  $focus-placeholder-z-index: calc($base-z-index + 30);
-
-  $placeholder-top: -12px;
-
-  @include use-lg-margin-y();
-
-  // inputのスタイル
-  .input {
-    height: 100%;
-    top: 50%;
-    transform: translateY(-50%);
-    -webkit-transform: translateY(-50%);
-    -ms-transform: translateY(-50%);
-    z-index: $input-z-index;
-    border: $blur-outline-color solid 1px;
-    border-radius: 4px;
-  }
-
-  // focusされたときのinputのスタイル
-  .input:focus {
-    border: $focus-outline-color solid 1px;
-    outline: none;
-    border-radius: 4px;
-  }
-
-  // 入力されていないときのplaceholderのスタイル
-  .input:placeholder-shown ~ .placeholder {
-    z-index: $blur-placeholder-z-index;
-  }
-
-  // focusされているときのplaceholderのスタイル
-  .input:focus ~ .placeholder {
-    top: $placeholder-top;
-    z-index: $focus-placeholder-z-index;
-  }
-
-  // 入力があるときのplaceholderのスタイル
-  .placeholder:not(:placeholder-shown) {
-    top: $placeholder-top;
-    z-index: $focus-placeholder-z-index;
-  }
-
-  // 入力がないときのplaceholder-backgroundのスタイル
-  .input:placeholder-shown ~ .placeholder-background {
-    position: absolute;
-    background-color: transparent;
-    color: transparent;
-    transition: all $b-transition-time;
-    transform-origin: 0;
-    transform: scale(0.75);
-    top: $placeholder-top;
-    left: 0;
-    margin: 0 10px;
-    padding: 0 2px;
-    z-index: $placeholder-background-z-index;
-  }
-
-  // 入力があるときのplaceholder-backgroundのスタイル
-  .placeholder-background {
-    position: absolute;
-    color: transparent;
-    transition: all $f-transition-time;
-    transform-origin: 0;
-    transform: scale(0.75);
-    top: $placeholder-top;
-    left: 0;
-    margin: 0 10px;
-    padding: 0 2px;
-    z-index: $placeholder-background-z-index;
-  }
-
-  // focusされているときのplaceholder-backgroundのスタイル
-  .input:focus ~ .placeholder-background {
-    background-color: white;
-  }
-
-  // 入力があるときのplaceholder-backgroundのスタイル
-  .input:not(:placeholder-shown) ~ .placeholder-background {
-    background-color: white;
-  }
-}
