--- packages/ui/src/components/server/text-field/variant/shared/use-focus-underline.scss
@@ -1,28 +1,0 @@
-@use '@theme/css/color/color.scss' as color;
-@use '@theme/css/font/index.scss' as font;
-
-$blur-outline-color: color.$accents4;
-$focus-outline-color: color.$suoh;
-$focus-color: color.$suoh;
-$blur-color: color.$accents3;
-$transition-time: 0.3s;
-
-@mixin use-focus-underline($z-index: 0) {
-  &::after {
-    border-bottom: 1px solid color.$suoh;
-    left: 0px;
-    bottom: 0px;
-    content: '';
-    position: absolute;
-    right: 0px;
-    transition: transform 200ms cubic-bezier(0, 0, 0.2, 1) 0ms;
-    pointer-events: none;
-  }
-  &:has(.input:focus)::after {
-    transform: scaleX(1);
-  }
-
-  &:not(:has(.input:focus))::after {
-    transform: scaleX(0);
-  }
-}
