--- packages/ui/src/components/server/text-field/variant/standard-text-field/standard-text-field.module.scss
@@ -1,38 +1,0 @@
-@use '@theme/css/color/color.scss' as color;
-@use '@theme/css/font/index.scss' as font;
-@use '@theme/css/margin/margin.scss' as *;
-@use '@layer/layer.scss' as layer;
-@use '../shared/use-focus-underline.scss' as *;
-
-/* anatomy
-.placeholder
-.input
-*/
-
-@mixin use-standard-text-field() {
-  $blur-outline-color: color.$accents4;
-  $focus-outline-color: color.$suoh;
-  $focus-color: color.$suoh;
-  $blur-color: color.$accents3;
-  $transition-time: 0.3s;
-  $base-z-index: layer.$earth;
-
-  @include use-focus-underline();
-  @include use-lg-margin-y();
-
-  .input {
-    height: 70%;
-    border-bottom: $blur-outline-color solid 1px;
-    bottom: 0;
-    z-index: $base-z-index;
-  }
-
-  .input:focus {
-    border-bottom: $blur-outline-color solid 1px;
-  }
-
-  .input:focus + .placeholder,
-  .input:not(:placeholder-shown) + .placeholder {
-    top: 2px;
-  }
-}
